+++ package.json
@@ -10,6 +10,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "console.table": "^0.10.0",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "inquirer": "^8.2.4",
